Separate out Mapping Filter Section to a component
