+++ src/app/integration/main/mapping/employee-mapping/employee-mapping.component.html
@@ -3,17 +3,7 @@
 
   <div *ngIf="!isLoading" [formGroup]="form" class="mapping--core-section">
     <div class="mapping--core-section-content">
-      <div fxLayout="row" fxLayoutAlign="start center" class="mapping--filters-section">
-        <div>
-          <app-simple-search-select [form]="form" [placeholder]="'Search Employee Email'" [showBackgroundColor]="false"></app-simple-search-select>
-        </div>
-        <div fxLayout="row" class="align-center">
-          <div [ngClass]="{'mapping--filter-all-selected': form.value.filterOption.length === filterOptions.length}">
-            <h5 class="mapping--filter-all-text pointer" (click)="addAllFilterHandler()">All</h5>
-          </div>
-          <h5 *ngFor="let option of filterOptions" [ngClass]="{'mapping--filter-deselected-text': form.value.filterOption.indexOf(option) === -1}" class="mapping--filter-alphabet-list pointer" (click)="filterOptionUpdateHandler(option)">{{option}}</h5>
-        </div>
-      </div>
+      <app-mapping-filter [form]="form" (mappingFilterUpdateHandler)="getMappings()"></app-mapping-filter>
 
       <div class="mapping--table">
         <table mat-table [dataSource]="mappings">
