+++ src/app/integration/main/mapping/employee-mapping/employee-mapping.component.scss
@@ -21,30 +21,6 @@
     }
   }
 
-  &--filters-section {
-   padding: 32px 0px 28px 24px;
-  }
-
-  &--filter-all-selected {
-    width: 31px;
-    height: 24px;
-    background: #F5F5F5;
-    border-radius: 12px;
-  }
-
-  &--filter-all-text {
-    padding: 4px;
-  }
-
-  &--filter-deselected-text {
-    color: #7b7c85;
-  }
-
-  &--filter-alphabet-list {
-    padding-left: 14px;
-    padding-top: 4px;
-  }
-
   &--heading {
     background: #F5F5F5;
     border: 1px solid #DFDFE2;
