+++ src/app/shared/components/mapping/mapping-filter/mapping-filter.component.html
@@ -1,1 +1,11 @@
-<p>mapping-filter works!</p>
+<div fxLayout="row" fxLayoutAlign="start center" class="mapping-filter--filters-section">
+  <div>
+    <app-simple-search-select [form]="form" [placeholder]="'Search Employee Email'" [showBackgroundColor]="false"></app-simple-search-select>
+  </div>
+  <div fxLayout="row" class="align-center">
+    <div [ngClass]="{'mapping-filter--filter-all-selected': form.value.filterOption.length === filterOptions.length}">
+      <h5 class="mapping-filter--filter-all-text pointer" (click)="addAllFilterHandler()">All</h5>
+    </div>
+    <h5 *ngFor="let option of filterOptions" [ngClass]="{'mapping-filter--filter-deselected-text': form.value.filterOption.indexOf(option) === -1}" class="mapping-filter--filter-alphabet-list pointer" (click)="filterOptionUpdateHandler(option)">{{option}}</h5>
+  </div>
+</div>
