+++ src/app/shared/components/mapping/mapping-filter/mapping-filter.component.scss
@@ -1,0 +1,25 @@
+.mapping-filter {
+  &--filters-section {
+    padding: 32px 0px 28px 24px;
+   }
+
+   &--filter-all-selected {
+     width: 31px;
+     height: 24px;
+     background: #F5F5F5;
+     border-radius: 12px;
+   }
+
+   &--filter-all-text {
+     padding: 4px;
+   }
+
+   &--filter-deselected-text {
+     color: #7b7c85;
+   }
+
+   &--filter-alphabet-list {
+     padding-left: 14px;
+     padding-top: 4px;
+   }
+}
