+++ src/app/shared/components/mapping/mapping-filter/mapping-filter.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-mapping-filter',
@@ -7,7 +8,29 @@ import { Component, OnInit } from '@angular/core';
 })
 export class MappingFilterComponent implements OnInit {
 
+  @Input() form: FormGroup;
+  filterOptions: string[] = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'];
+  @Output() mappingFilterUpdateHandler = new EventEmitter<{}>();
+
   constructor() { }
+
+  addAllFilterHandler(): void {
+    this.form.controls.filterOption.patchValue(this.filterOptions.concat());
+
+    this.mappingFilterUpdateHandler.emit({});
+  }
+
+  filterOptionUpdateHandler(alphabet: string): void {
+    const index = this.form.value.filterOption.indexOf(alphabet);
+
+    if (index > -1) {
+      this.form.value.filterOption.splice(index, 1);
+    } else {
+      this.form.value.filterOption.push(alphabet);
+    }
+
+    this.mappingFilterUpdateHandler.emit({});
+  }
 
   ngOnInit(): void {
   }
